requirements.txt
langchain
python-dotenv 
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
sentence-transformers
langchain_ollama
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_core
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette